+++ crates/jsonschema-py/tests-py/test_evaluate.py
@@ -39,7 +39,19 @@ def test_evaluate_produces_expected_outputs_for_valid_instance():
                 "evaluationPath": "/properties",
                 "instanceLocation": "",
                 "schemaLocation": "/properties",
-                "annotations": ["age", "name"],
+                "annotations": ["name", "age"],
+            },
+            {
+                "valid": True,
+                "evaluationPath": "/properties/name",
+                "instanceLocation": "/name",
+                "schemaLocation": "/properties/name",
+            },
+            {
+                "valid": True,
+                "evaluationPath": "/properties/name/type",
+                "instanceLocation": "/name",
+                "schemaLocation": "/properties/name/type",
             },
             {
                 "valid": True,
@@ -58,18 +70,6 @@ def test_evaluate_produces_expected_outputs_for_valid_instance():
                 "evaluationPath": "/properties/age/minimum",
                 "instanceLocation": "/age",
                 "schemaLocation": "/properties/age/minimum",
-            },
-            {
-                "valid": True,
-                "evaluationPath": "/properties/name",
-                "instanceLocation": "/name",
-                "schemaLocation": "/properties/name",
-            },
-            {
-                "valid": True,
-                "evaluationPath": "/properties/name/type",
-                "instanceLocation": "/name",
-                "schemaLocation": "/properties/name/type",
             },
         ],
     }
@@ -97,8 +97,22 @@ def test_evaluate_produces_expected_outputs_for_valid_instance():
                 "evaluationPath": "/properties",
                 "instanceLocation": "",
                 "schemaLocation": "/properties",
-                "annotations": ["age", "name"],
+                "annotations": ["name", "age"],
                 "details": [
+                    {
+                        "valid": True,
+                        "evaluationPath": "/properties/name",
+                        "instanceLocation": "/name",
+                        "schemaLocation": "/properties/name",
+                        "details": [
+                            {
+                                "valid": True,
+                                "evaluationPath": "/properties/name/type",
+                                "instanceLocation": "/name",
+                                "schemaLocation": "/properties/name/type",
+                            }
+                        ],
+                    },
                     {
                         "valid": True,
                         "evaluationPath": "/properties/age",
@@ -119,20 +133,6 @@ def test_evaluate_produces_expected_outputs_for_valid_instance():
                             },
                         ],
                     },
-                    {
-                        "valid": True,
-                        "evaluationPath": "/properties/name",
-                        "instanceLocation": "/name",
-                        "schemaLocation": "/properties/name",
-                        "details": [
-                            {
-                                "valid": True,
-                                "evaluationPath": "/properties/name/type",
-                                "instanceLocation": "/name",
-                                "schemaLocation": "/properties/name/type",
-                            }
-                        ],
-                    },
                 ],
             },
         ],
@@ -143,7 +143,7 @@ def test_evaluate_produces_expected_outputs_for_valid_instance():
             "schemaLocation": "/properties",
             "absoluteKeywordLocation": None,
             "instanceLocation": "",
-            "annotations": ["age", "name"],
+            "annotations": ["name", "age"],
         }
     ]
     assert evaluation.errors() == []